+++ core/src/main/java/com/graphhopper/routing/weighting/Weighting.java
@@ -43,7 +43,8 @@ public interface Weighting {
      * @param reverse   if the specified edge is specified in reverse direction e.g. from the reverse
      *                  case of a bidirectional search.
      * @return the calculated weight with the specified velocity has to be in the range of 0 and
-     * +Infinity. Make sure your method does not return NaN which can e.g. occur for 0/0.
+     * +Infinity. GraphHopper expects weights to be whole numbers only. Consider using {@link Weighting#roundWeight(double)}
+     * to post-process all weights. Make sure your method does not return NaN which can e.g. occur for 0/0.
      */
     double calcEdgeWeight(EdgeIteratorState edgeState, boolean reverse);