+++ core/src/test/java/com/graphhopper/routing/RoutingAlgorithmTest.java
@@ -19,6 +19,7 @@ package com.graphhopper.routing;
 
 import com.carrotsearch.hppc.IntArrayList;
 import com.carrotsearch.hppc.IntIndexedContainer;
+import com.graphhopper.json.Statement;
 import com.graphhopper.routing.ch.CHRoutingAlgorithmFactory;
 import com.graphhopper.routing.ch.PrepareContractionHierarchies;
 import com.graphhopper.routing.ev.DecimalEncodedValue;
@@ -32,6 +33,8 @@ import com.graphhopper.routing.util.TraversalMode;
 import com.graphhopper.routing.weighting.SpeedWeighting;
 import com.graphhopper.routing.weighting.TurnCostProvider;
 import com.graphhopper.routing.weighting.Weighting;
+import com.graphhopper.routing.weighting.custom.CustomModelParser;
+import com.graphhopper.routing.weighting.custom.CustomWeighting;
 import com.graphhopper.storage.*;
 import com.graphhopper.storage.index.LocationIndex;
 import com.graphhopper.storage.index.LocationIndexTree;
@@ -44,6 +47,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.ArgumentsProvider;
 import org.junit.jupiter.params.provider.ArgumentsSource;
+import org.locationtech.jts.geom.Coordinate;
+import org.locationtech.jts.geom.GeometryFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -865,6 +870,52 @@ public class RoutingAlgorithmTest {
         p = f.calcPath(graph, new GHPoint(0.00009, 0.00011), new GHPoint(0.00001, 0.00011));
         assertEquals(nodes(6, 2, 5), p.calcNodes(), p.toString());
         assertEquals(12.57, p.getDistance(), .1, p.toString());
+    }
+
+    @ParameterizedTest
+    @ArgumentsSource(FixtureProvider.class)
+    public void testBlockArea(Fixture f) {
+        // 1 -y- 2
+        // |  5  |
+        // x 4   |  the area around node 1 is expensive (node 4 is outside this area)
+        // |     |
+        // 0 --- 3
+        BaseGraph graph = f.createGHStorage();
+        graph.edge(0, 1).set(f.carSpeedEnc, 10, 10);
+        graph.edge(1, 2).set(f.carSpeedEnc, 10, 10);
+        graph.edge(2, 3).set(f.carSpeedEnc, 10, 10);
+        graph.edge(3, 0).set(f.carSpeedEnc, 10, 10);
+        updateDistancesFor(graph, 0, 40.000, 10.000);
+        updateDistancesFor(graph, 1, 40.001, 10.000);
+        updateDistancesFor(graph, 2, 40.001, 10.001);
+        updateDistancesFor(graph, 3, 40.000, 10.001);
+
+        JsonFeatureCollection areas = new JsonFeatureCollection();
+        Coordinate[] blockArea = new Coordinate[]{
+                new Coordinate(9.9997, 40.0007),
+                new Coordinate(9.9997, 40.0013),
+                new Coordinate(10.0003, 40.0013),
+                new Coordinate(10.0003, 40.0007),
+                new Coordinate(9.9997, 40.0007)
+        };
+        areas.getFeatures().add(new JsonFeature("expensive",
+                "Feature",
+                null,
+                new GeometryFactory().createPolygon(blockArea),
+                new HashMap<>()));
+        CustomModel customModel = new CustomModel()
+                .addToSpeed(Statement.If("true", Statement.Op.LIMIT, "10"))
+                .addToPriority(Statement.If("in_expensive", Statement.Op.MULTIPLY, "0.1"));
+        customModel.addAreas(areas);
+
+        CustomWeighting weighting = CustomModelParser.createWeighting(f.encodingManager, TurnCostProvider.NO_TURN_COST_PROVIDER, customModel);
+
+        // We route from x(4) to y(5). The virtual edges adjacent to node 1 are in the area and thus expensive,
+        // so we need to take the detour via 3. But x-0 and y-2 are not. If QueryOverlay#adjustWeights distributed
+        // the weight difference between the original edges 0-1 and 1-2 (penalized all the way) equally to the
+        // virtual edges also x-0 and y-2 would carry the penalty and the route would go through the expensive area.
+        Path path = f.calcPath(graph, weighting, new GHPoint(40.0006, 10.000), new GHPoint(40.001, 10.0004));
+        assertEquals(nodes(4, 0, 3, 2, 5), path.calcNodes());
     }
 
     @ParameterizedTest